Stop using `Docker`; Use built in `webpack` compression.
This redacts the `Docker` based mechanism for sharing code
I previously derived. We plan on sharing code by storing a static
version in the repository itself (but including instructions for
building and updating).

This also adds support for a `build` target that uses the built
in `webpack` compression by setting the `mode` when `NODE_ENV`
indicates we're building for production.